Update Rewards to display in User Wallet Summary
