+++ src/locales/en.json
@@ -146,6 +146,7 @@
   "editor": "Editor",
   "preview": "Preview",
   "reward": "Reward",
+  "rewards": "Rewards",
   "reward_option_100": "100% Steem Power",
   "reward_option_50": "50% SBD and 50% SP",
   "reward_option_0": "Declined",
@@ -197,7 +198,6 @@
   "cancel_transfer_from_savings": "Cancel transfer from savings (request {requestId})",
   "received_from": "Received from {username}",
   "claim_rewards": "Claim Rewards ",
-  "your_current_rewards": "Your current rewards",
   "reward_claimed": "Reward Claimed",
   "powered_up": "Powered up ",
   "market": "Market"
