--- src/wallet/ClaimRewardsButton.less
@@ -1,21 +1,0 @@
-@import (reference) "../styles/custom.less";
-
-.ClaimRewardsButton {
-  &__title {
-    color: @black;
-    font-weight: 500;
-  }
-
-  &__reward {
-    &:after {
-      color: @blue-botticelli;
-      content: ' - ';
-    }
-
-    &:last-of-type {
-      &:after {
-        content: '';
-      }
-    }
-  }
-}
